Fix review feedback: bigint cast, phoneticScript normalization, no-JS form fields
- Cast count(*) OVER() to ::int in common-components query to avoid
  postgres.js returning bigint as string
- Normalize 'pinyin' to null in settings action (null = default pinyin)
- Wrap hidden JSON inputs in @media (scripting: enabled) guard so they
  don't shadow noscript inputs when JS is disabled
- Add same scripting guard to key-value-editor's .js-only class

+++ src/lib/server/services/dictionary.ts
@@ -219,7 +219,7 @@ export async function getCharacterList(
 					c.jun_da_per_million AS "junDaPerMillion",
 					c.simplified_variants AS "simplifiedVariants",
 					c.traditional_variants AS "traditionalVariants",
-					count(*) OVER() AS total
+					count(*) OVER()::int AS total
 				FROM component_counts cc
 				JOIN ${charView} c ON c.character = cc.character
 				ORDER BY cc.usage_count DESC
